hipcc: fix default include path on Windows and adapt to TheRock rocm layout
getCppConfig in the orignial code did not add the content of
`getHipPath()` to the default C Preprocessor definitions in the
`if (isWindows())` path.

Change default `hipClangPath` to look for clang under `lib/llvm/bin`
like Linux instead of expecting `clang.exe` to be next to `hipcc.exe`.

Change output of `hipcc.exe --ldflags` to use Windows "\" path
separators to make the output obviously correct.