[Flang] Fix defaultmap(none) being overly aggressive with symbol checks
Currently we're picking up and complaining about builtin symbols like
Null() when defaultmap(none) is set, so I've relaxed the restriction a
bit to allow for procedures and named constants to bypass the
restriction. It might be the case that we want to tighten it up again
in certain aspects in the future.